move data to json file
