+++ Web/starter_project_iconics/data/blog/blog-tile.json
@@ -1,0 +1,32 @@
+[
+  {
+    "name": "Yididiya Kebede",
+    "profession": "SOFTWARE ENGINEER",
+    "profileImg": "/img/blog/profile-img.jpg",
+    "title": "The essential guide to Competitive Programming Tab System On React : 3 ways to do it.",
+    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea ",
+    "datePosted": "Apr 16, 2022",
+    "tags": ["UI/UX", "Development"],
+    "blogImg": "/img/blog/blog-img.jpg"
+  },
+  {
+    "name": "Yididiya Kebede",
+    "profession": "SOFTWARE ENGINEER",
+    "profileImg": "/img/blog/profile-img.jpg",
+    "title": "The essential guide to Competitive Programming Tab System On React : 3 ways to do it.",
+    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea ",
+    "datePosted": "Apr 16, 2022",
+    "tags": ["UI/UX", "Development"],
+    "blogImg": "/img/blog/blog-img.jpg"
+  },
+  {
+    "name": "Yididiya Kebede",
+    "profession": "SOFTWARE ENGINEER",
+    "profileImg": "/img/blog/profile-img.jpg",
+    "title": "The essential guide to Competitive Programming Tab System On React : 3 ways to do it.",
+    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea ",
+    "datePosted": "Apr 16, 2022",
+    "tags": ["UI/UX", "Development"],
+    "blogImg": "/img/blog/blog-img.jpg"
+  }
+]
