+++ Web/starter_project_iconics/types/blog/blog.tsx
@@ -2,7 +2,7 @@ export interface BlogData {
   name: string
   profileImg: string
   profession: string
-  datePosted: Date
+  datePosted: string
   title: String
   description: string
   tags: string[]
